deep-variational-autoencoder-MNIST-Kaggle
This notebook constructs a deep variational autoencoder (VAE) using the Keras package in python and applies it to the MNIST data in the context of Kaggle's Digit Recognizer competition (so some labels are missing). Some code is borrowed from @fchollet's example from his book 'Deep Learning with Python'. See the book repo for more details.